requirements.txt
torch
pillow
ipython
tqdm
pymilvus
python-dotenv
streamlit
pyyaml
colorlog
sentence-transformers
